+++ src/zope/server/http/httprequestparser.py
@@ -39,17 +39,31 @@ class HTTPRequestParser(object):
     a server task constructor.
     """
 
+    # Parsing status
     completed = 0  # Set once request is completed.
     empty = 0        # Set if no request was made.
+    in_header = False
+
     header_plus = b''
     chunked = 0
     content_length = 0
     body_rcv = None
-    # Other attributes: first_line, header, headers, command, uri, version,
-    # path, query, fragment
 
-    # headers is a mapping containing keys translated to uppercase
-    # with dashes turned into underscores.
+    # Data from parsing. native strings.
+    first_line = ''
+    header = ''
+    command = ''
+    uri = ''
+    version = ''
+    proxy_scheme = None
+    proxy_netloc = None
+    path = None
+    fragment = None
+    query = None
+
+    # headers is a mapping containing native string keys translated to
+    # uppercase with dashes turned into underscores. The values
+    # are also native strings.
 
     def __init__(self, adj):
         """
@@ -135,9 +149,10 @@ class HTTPRequestParser(object):
                     headers[key1] = value
             # else there's garbage in the headers?
 
+        assert isinstance(self.first_line, str)
         command, uri, version = self.crack_first_line()
-        self.command = str(command)
-        self.uri = str(uri)
+        self.command = command or ''
+        self.uri = uri or ''
         self.version = version
         self.split_uri()