+++ src/zope/server/http/httptask.py
@@ -16,8 +16,6 @@
 An HTTP task that can execute an HTTP request with the help of the channel and
 the server it belongs to.
 """
-import socket
-import time
 
 from zope.server.http.http_date import build_http_date
 from zope.publisher.interfaces.http import IHeaderOutput
@@ -50,6 +48,7 @@ class HTTPTask(AbstractTask):
     cgi_env = None
 
     def __init__(self, channel, request_data):
+        # request_data is a httprequestparser.HTTPRequestParser
         AbstractTask.__init__(self, channel)
         self.request_data = request_data
         self.response_headers = {}
@@ -98,7 +97,7 @@ class HTTPTask(AbstractTask):
 
         if version == '1.0':
             if connection == 'keep-alive':
-                if not ('Content-Length' in response_headers):
+                if 'Content-Length' not in response_headers:
                     close_it = 1
                 else:
                     response_headers['Connection'] = 'Keep-Alive'
@@ -106,21 +105,21 @@ class HTTPTask(AbstractTask):
                 close_it = 1
         elif version == '1.1':
             if 'connection: close' in (header.lower() for header in
-                accumulated_headers):
+                                       accumulated_headers):
                 close_it = 1
             if connection == 'close':
                 close_it = 1
             elif 'Transfer-Encoding' in response_headers:
-                if not response_headers['Transfer-Encoding'] == 'chunked':
+                if response_headers['Transfer-Encoding'] != 'chunked':
                     close_it = 1
             elif self.status == '304':
                 # Replying with headers only.
                 pass
-            elif not ('Content-Length' in response_headers):
+            elif 'Content-Length' not in response_headers:
                 # accumulated_headers is a simple list, we need to cut off
                 # the value of content-length manually
                 if 'content-length' not in (header[:14].lower() for header in
-                    accumulated_headers):
+                                            accumulated_headers):
                     close_it = 1
             # under HTTP 1.1 keep-alive is default, no need to set the header
         else:
@@ -139,7 +138,7 @@ class HTTPTask(AbstractTask):
         else:
             self.response_headers['Via'] = self.channel.server.SERVER_IDENT
         if 'date' not in (header[:4].lower() for header in
-                            accumulated_headers):
+                          accumulated_headers):
             self.response_headers['Date'] = build_http_date(self.start_time)