+++ CHANGES.rst
@@ -11,6 +11,10 @@
 
 - Add PyPy support.
 
+- Made the HTTPTask not have ``command`` or ``uri`` values of
+  ``"None"`` when the first request line cannot be parsed. Now they
+  are empty strings.
+
 3.9.0 (2013-03-13)
 ==================